@@ -1,5 +1,5 @@
 [metadata]
-name = lnPi
+name = lnpy
 author = William P. Krekelberg
 author_email = wpk@nist.gov
 license= NIST license
@@ -9,7 +9,7 @@ description = lnPi analysis
 # long_description_content_type = text/markdown
 long_description = file: README.rst, HISTORY.rst, LICENSE
 keywords = TMMC
-url = https://github.com/wpk-nist-gov/lnPi
+url = https://github.com/wpk-nist-gov/lnpy
 classifiers =
     Development Status :: 2 - Pre-Alpha
     License :: Public Domain
@@ -17,10 +17,9 @@ classifiers =
     Intended Audience :: Science/Research
     Programming Language :: Python
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.6
-    Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
     Topic :: Scientific/Engineering
 
 
@@ -29,21 +28,25 @@ classifiers =
 packages = find:
 zip_safe = True # if using mypy, must be False
 include_package_data = True
-python_requires = >=3.6
+python_requires = >=3.7
 install_requires =
-    setuptools >= 38.4
     numpy
     xarray
     scipy
     scikit-image
     joblib
-    dask
-    netcdf4
     bottleneck
-setup_requires =
-    setuptools >= 38.4
-    setuptools_scm >= 5.0
+    tqdm
+    importlib_resources
+    # dask
+    # netcdf4
 
+
+
+
+[options.package_data]
+xarray =
+    lnpy/data/*
 
 [aliases]
 test = pytest
@@ -58,7 +61,7 @@ profile = black
 skip_gitignore = true
 force_to_top = true
 default_section = THIRDPARTY
-known_first_party = lnPi
+known_first_party = lnpy
 
 [flake8]
 ignore =
