@@ -1,21 +1,21 @@
-from __future__ import print_function, absolute_import, division
+from __future__ import absolute_import, division, print_function
 
+from warnings import warn
 
 import numpy as np
 import pandas as pd
 from scipy.ndimage import filters
 
 from .cached_decorators import gcached
-from .utils import labels_to_masks, masks_change_convention
-
 from .extensions import AccessorMixin
+from .utils import labels_to_masks, masks_change_convention
 
 # NOTE : This is a rework of core.
 # [ ] : split xarray functionality into wrapper(s)
 # [ ] : split splitting into separate classes
 
 
-class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
+class MaskedlnPiLegacy(np.ma.MaskedArray, AccessorMixin):
     """
     class to store masked ln[Pi(n0,n1,...)].
     shape is (N0,N1,...) where Ni is the span of each dimension)
@@ -34,12 +34,8 @@ class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
     reweight : create new lnPi at new mu
     smooth : create smoothed object
     """
-    def __new__(cls,
-                data=None,
-                lnz=None,
-                state_kws=None,
-                extra_kws=None,
-                **kwargs):
+
+    def __new__(cls, data=None, lnz=None, state_kws=None, extra_kws=None, **kwargs):
         """
         constructor
 
@@ -64,8 +60,10 @@ class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
         kwargs : arguments to np.ma.array
             e.g., mask=...
         """
+        warn("MaskedlnPiLegacy is deprecated.  Please use lnPiMasked instead")
+
         if data is not None and issubclass(data.dtype.type, np.floating):
-            kwargs.setdefault('fill_value', np.nan)
+            kwargs.setdefault("fill_value", np.nan)
 
         obj = np.ma.array(data, **kwargs).view(cls)
         # fv = kwargs.get('fill_value', None) or getattr(data, 'fill_value', None)
@@ -82,7 +80,7 @@ class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
             lnz = np.zeros(obj.ndim)
         lnz = np.atleast_1d(lnz).astype(obj.dtype)
         if len(lnz) != obj.ndim:
-            raise ValueError('bad len on lnz %s' % lnz)
+            raise ValueError("bad len on lnz %s" % lnz)
 
         if state_kws is None:
             state_kws = {}
@@ -97,7 +95,7 @@ class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
         return obj
 
     ##################################################
-    #caching
+    # caching
     def __array_finalize__(self, obj):
         super().__array_finalize__(obj)
         self._clear_cache()
@@ -106,7 +104,7 @@ class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
         self._cache = {}
 
     ##################################################
-    #properties
+    # properties
     @property
     def optinfo(self):
         """all extra properties"""
@@ -115,19 +113,19 @@ class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
     @property
     def state_kws(self):
         """state specific parameters"""
-        return self._optinfo['state_kws']
+        return self._optinfo["state_kws"]
 
     @property
     def extra_kws(self):
         """all extra parameters"""
-        return self._optinfo['extra_kws']
+        return self._optinfo["extra_kws"]
 
     def _index_dict(self, phase=None):
 
-        out = {'lnz_{}'.format(i): v for i, v in enumerate(self.lnz)}
+        out = {"lnz_{}".format(i): v for i, v in enumerate(self.lnz)}
         if phase is not None:
-            out['phase'] = phase
-        #out.update(**self.state_kws)
+            out["phase"] = phase
+        # out.update(**self.state_kws)
         return out
 
     def _lnpi_tot(self, fill_value=None):
@@ -153,10 +151,9 @@ class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
     # def _lnpi_0_tot(self):
     #     return self.data.ravel()[0]
 
-
     @property
     def lnz(self):
-        return self._optinfo.get('lnz', None)
+        return self._optinfo.get("lnz", None)
 
     @property
     def betamu(self):
@@ -167,38 +164,38 @@ class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
     #     return self._optinfo.get('mu', None)
     @property
     def volume(self):
-        return self.state_kws.get('volume', None)
+        return self.state_kws.get("volume", None)
 
     @property
     def beta(self):
-        return self.state_kws.get('beta', None)
+        return self.state_kws.get("beta", None)
 
     def __repr__(self):
         L = []
-        L.append('lnz={}'.format(repr(self.lnz)))
-        L.append('state_kws={}'.format(repr(self.state_kws)))
+        L.append("lnz={}".format(repr(self.lnz)))
+        L.append("state_kws={}".format(repr(self.state_kws)))
 
-        L.append('data={}'.format(super(MaskedlnPi, self).__repr__()))
+        L.append("data={}".format(super().__repr__()))
         if len(self.extra_kws) > 0:
-            L.append('extra_kws={}'.format(repr(self.extra_kws)))
+            L.append("extra_kws={}".format(repr(self.extra_kws)))
 
-        indent = ' ' * 5
-        p = 'MaskedlnPi(\n' + '\n'.join([indent + x for x in L]) + '\n)'
+        indent = " " * 5
+        p = "MaskedlnPi(\n" + "\n".join([indent + x for x in L]) + "\n)"
 
         return p
 
     def __str__(self):
-        return 'MaskedlnPi(lnz={})'.format(str(self.lnz))
+        return "MaskedlnPi(lnz={})".format(str(self.lnz))
 
-    #@gcached(prop=False)
+    # @gcached(prop=False)
     def local_argmax(self, *args, **kwargs):
         return np.unravel_index(self.argmax(*args, **kwargs), self.shape)
 
-    #@gcached(prop=False)
+    # @gcached(prop=False)
     def local_max(self, *args, **kwargs):
         return self[self.local_argmax(*args, **kwargs)]
 
-    #@gcached(prop=False)
+    # @gcached(prop=False)
     def local_maxmask(self, *args, **kwargs):
         return self == self.local_max(*args, **kwargs)
 
@@ -206,13 +203,14 @@ class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
     def edge_distance_matrix(self):
         """matrix of distance from upper bound"""
         from .utils import distance_matrix
+
         return distance_matrix(~self.mask)
 
     def edge_distance(self, ref, *args, **kwargs):
         return ref.edge_distance_matrix[self.local_argmax(*args, **kwargs)]
 
     # make these top level
-    #@gcached()
+    # @gcached()
     # @property
     # def pi(self):
     #     """
@@ -236,12 +234,7 @@ class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
         self._clear_cache()
         super().__setitem__(index, value)
 
-    def pad(self,
-            axes=None,
-            ffill=True,
-            bfill=False,
-            limit=None,
-            inplace=False):
+    def pad(self, axes=None, ffill=True, bfill=False, limit=None, inplace=False):
         """
         pad nan values in underlying data to values
 
@@ -263,8 +256,9 @@ class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
         out : lnPi
             padded object
         """
-        from .utils import ffill, bfill
         import bottleneck
+
+        from .utils import bfill, ffill
 
         if axes is None:
             axes = range(self.ndim)
@@ -342,16 +336,16 @@ class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
 
         lnz = np.atleast_1d(lnz)
 
-        assert (len(lnz) == len(self.lnz))
+        assert len(lnz) == len(self.lnz)
 
         new = self.copy()
-        new._optinfo['lnz'] = lnz
+        new._optinfo["lnz"] = lnz
 
         dlnz = new.lnz - self.lnz
 
-        #s = _get_shift(self.shape,dmu)*self.beta
-        #get shift
-        #i.e., N * (mu_1 - mu_0)
+        # s = _get_shift(self.shape,dmu)*self.beta
+        # get shift
+        # i.e., N * (mu_1 - mu_0)
         # note that this is (for some reason)
         # faster than doing the (more natural) options:
         # N = self.ncoords.values
@@ -365,20 +359,15 @@ class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
         for i, (s, m) in enumerate(zip(self.shape, dlnz)):
             shift = np.add.outer(shift, np.arange(s) * m)
 
-        #scale by beta
-        #shift *= self.beta
+        # scale by beta
+        # shift *= self.beta
 
         new.data[...] += shift
         new.adjust(zeromax=zeromax, pad=pad, inplace=True)
 
         return new
 
-    def smooth(self,
-               sigma=4,
-               mode='nearest',
-               truncate=4,
-               inplace=False,
-               **kwargs):
+    def smooth(self, sigma=4, mode="nearest", truncate=4, inplace=False, **kwargs):
         """
         apply gaussian filter smoothing to data
 
@@ -398,12 +387,14 @@ class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
         else:
             new = self.copy()
 
-        filters.gaussian_filter(new.data,
-                                output=new.data,
-                                mode=mode,
-                                truncate=truncate,
-                                sigma=sigma,
-                                **kwargs)
+        filters.gaussian_filter(
+            new.data,
+            output=new.data,
+            mode=mode,
+            truncate=truncate,
+            sigma=sigma,
+            **kwargs
+        )
         return new
 
     def copy_shallow(self, mask=None, **kwargs):
@@ -419,10 +410,12 @@ class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
         if mask is None:
             mask = self.mask.copy()
 
-        return self.__class__(self.data,
-                              mask=mask,
-                              fill_value=self.fill_value,
-                              **dict(self._optinfo, **kwargs))
+        return self.__class__(
+            self.data,
+            mask=mask,
+            fill_value=self.fill_value,
+            **dict(self._optinfo, **kwargs)
+        )
 
     def or_mask(self, mask, **kwargs):
         """
@@ -443,29 +436,23 @@ class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
 
     def __setstate__(self, state):
         ma, opt = state
-        super(MaskedlnPi, self).__setstate__(ma)
+        super().__setstate__(ma)
         self._optinfo.update(opt)
 
-
-#        opt = self._optinfo
-#        return ma, opt
-#         # ma = self.view(np.ma.MaskedArray)
-#         # info = self._optinfo
-#         # return ma, info
-#self._optinfo.update(opt)
-# ma, info = state
-#         # super(MaskedlnPi, self).__setstate__(ma)
-#         # self._optinfo.update(info)
+    #        opt = self._optinfo
+    #        return ma, opt
+    #         # ma = self.view(np.ma.MaskedArray)
+    #         # info = self._optinfo
+    #         # return ma, info
+    # self._optinfo.update(opt)
+    # ma, info = state
+    #         # super(MaskedlnPi, self).__setstate__(ma)
+    #         # self._optinfo.update(info)
 
     @classmethod
-    def from_table(cls,
-                   path,
-                   lnz,
-                   state_kws=None,
-                   sep='\s+',
-                   names=None,
-                   csv_kws=None,
-                   **kwargs):
+    def from_table(
+        cls, path, lnz, state_kws=None, sep=r"\s+", names=None, csv_kws=None, **kwargs
+    ):
         """
         Create lnPi object from text file table with columns [n_0,...,n_ndim, lnpi]
 
@@ -489,18 +476,23 @@ class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
         ndim = len(lnz)
 
         if names is None:
-            names = ['n_{}'.format(i) for i in range(ndim)] + ['lnpi']
+            names = ["n_{}".format(i) for i in range(ndim)] + ["lnpi"]
 
         if csv_kws is None:
             csv_kws = {}
 
-        da = (pd.read_csv(path, sep=sep, names=names,
-                          **csv_kws).set_index(names[:-1])['lnpi'].to_xarray())
-        return cls(data=da.values,
-                   mask=da.isnull().values,
-                   lnz=lnz,
-                   state_kws=state_kws,
-                   **kwargs)
+        da = (
+            pd.read_csv(path, sep=sep, names=names, **csv_kws)
+            .set_index(names[:-1])["lnpi"]
+            .to_xarray()
+        )
+        return cls(
+            data=da.values,
+            mask=da.isnull().values,
+            lnz=lnz,
+            state_kws=state_kws,
+            **kwargs
+        )
 
     @classmethod
     def from_dataarray(cls, da, state_as_attrs=None, **kwargs):
@@ -509,15 +501,15 @@ class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
         """
 
         kws = {}
-        kws['data'] = da.values
-        if 'mask' in da.coords:
-            kws['mask'] = da.mask.values
+        kws["data"] = da.values
+        if "mask" in da.coords:
+            kws["mask"] = da.mask.values
         else:
-            kws['mask'] = da.isnull().values
+            kws["mask"] = da.isnull().values
 
         # where are state variables
         if state_as_attrs is None:
-            state_as_attrs = bool(da.attrs.get('state_as_attrs', False))
+            state_as_attrs = bool(da.attrs.get("state_as_attrs", False))
         if state_as_attrs:
             # state variables from attrs
             c = da.attrs
@@ -526,20 +518,20 @@ class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
 
         lnz = []
         state_kws = {}
-        for k in da.attrs['dims_state']:
+        for k in da.attrs["dims_state"]:
             val = np.array(c[k])
-            if 'lnz' in k:
+            if "lnz" in k:
                 lnz.append(val)
             else:
                 state_kws[k] = val * 1
-        kws['lnz'] = lnz
-        kws['state_kws'] = state_kws
+        kws["lnz"] = lnz
+        kws["state_kws"] = state_kws
 
         # any overrides
         kwargs = dict(kws, **kwargs)
         return cls(**kwargs)
 
-    def list_from_masks(self, masks, convention='image'):
+    def list_from_masks(self, masks, convention="image"):
         """
         create list of lnpis corresponding to masks[i]
 
@@ -556,27 +548,27 @@ class MaskedlnPi(np.ma.MaskedArray, AccessorMixin):
         """
 
         return [
-            self.or_mask(m)
-            for m in masks_change_convention(masks, convention, False)
+            self.or_mask(m) for m in masks_change_convention(masks, convention, False)
         ]
 
-    def list_from_labels(self,
-                         labels,
-                         features=None,
-                         include_boundary=False,
-                         check_features=True,
-                         **kwargs):
+    def list_from_labels(
+        self,
+        labels,
+        features=None,
+        include_boundary=False,
+        check_features=True,
+        **kwargs
+    ):
         """
         create list of lnpis corresponding to labels
         """
 
-        masks, features = labels_to_masks(labels=labels,
-                                          features=features,
-                                          include_boundary=include_boundary,
-                                          convention=False,
-                                          check_features=check_features,
-                                          **kwargs)
+        masks, features = labels_to_masks(
+            labels=labels,
+            features=features,
+            include_boundary=include_boundary,
+            convention=False,
+            check_features=check_features,
+            **kwargs
+        )
         return self.list_from_masks(masks, convention=False)
-
-
-
