@@ -1,16 +1,26 @@
+import json
+from dataclasses import dataclass
+from typing import Any
+
+import numpy as np
+import pandas as pd
 import pytest
 
+import lnpy
+import lnpy.stability
 
-import lnPi, lnPi.stability
-import json
-import pandas as pd
-import numpy as np
+
+def tag_phases2(x):
+    if len(x) > 2:
+        raise ValueError("bad tag function")
+    argmax0 = np.array([xx.local_argmax()[0] for xx in x])
+    return np.where(argmax0 <= x[0].shape[0] / 2, 0, 1)
 
 
 from pathlib import Path
 
+path_data = Path(__file__).parent / "../examples/LJ_cfs_2.5sig"
 
-path_data = Path(__file__).parent / '../examples/LJ_cfs_2.5sig'
 
 def get_lnz(path):
     kB = 1.3806503e-23  # J/K
@@ -41,26 +51,22 @@ def get_lnz(path):
     return lnz, {"beta": 1.0 / temp, "volume": metadata["V*"]}
 
 
-def tag_phases2(x):
-    if len(x) > 2:
-        raise ValueError('bad tag function')
-    argmax0 = np.array([xx.local_argmax()[0] for xx in x])
-    return np.where(argmax0 <= x[0].shape[0] / 2, 0, 1)
-
-
 # get meta data
 @pytest.fixture
 def ref():
-    lnz, state_kws = get_lnz(path_data /"t072871.metadata.json")
+    lnz, state_kws = get_lnz(path_data / "t072871.metadata.json")
     print(state_kws)
 
     # read in potential energy
     pe = pd.read_csv(
-        path_data /"ljsf.t072871.bulk.v512.r1.energy.dat", header=None, sep="\s+", names=["n", "e"]
+        path_data / "ljsf.t072871.bulk.v512.r1.energy.dat",
+        header=None,
+        sep=r"\s+",
+        names=["n", "e"],
     )["e"].values
 
     return (
-        lnPi.MaskedlnPiDelayed.from_table(
+        lnpy.lnPiMasked.from_table(
             path_data / "ljsf.t072871.bulk.v512.r1.lnpi.dat",
             fill_value=np.nan,
             lnz=lnz,
@@ -78,7 +84,7 @@ def ref():
 
 @pytest.fixture
 def phase_creator(ref):
-    return lnPi.segment.PhaseCreator(
+    return lnpy.segment.PhaseCreator(
         nmax=2,
         nmax_peak=4,
         ref=ref,
@@ -89,7 +95,7 @@ def phase_creator(ref):
         # MaskedlnPi objects, and should return
         # a list of phaseid
         # tag_phases=None
-        tag_phases=tag_phases2
+        tag_phases=tag_phases2,
     )
 
 
@@ -99,6 +105,20 @@ def build_phases(phase_creator):
     return phase_creator.build_phases_mu([None])
 
 
+import lnpy.examples
+
+
+@pytest.fixture(params=[1])
+def obj(request, ref, phase_creator, build_phases):
+    if request.param == 0:
+        return lnpy.examples.Example(
+            ref=ref, phase_creator=phase_creator, build_phases=build_phases
+        )
+    else:
+
+        return lnpy.examples.lj_sub_example()
+
+
 @pytest.fixture
 def test_table():
     table = pd.read_csv(path_data / "data_1.csv")
@@ -106,18 +126,35 @@ def test_table():
 
 
 def get_test_table(o, ref):
-    return o.xge.table(keys=['betaOmega','nvec','PE','dens','betaF','S','betaG','edge_distance'], ref=ref).to_dataframe().reset_index()#.to_csv('data_0.csv', index=False)
+    return (
+        o.xge.table(
+            keys=[
+                "betaOmega",
+                "nvec",
+                "PE",
+                "dens",
+                "betaF",
+                "S",
+                "betaG",
+                "edge_distance",
+            ],
+            ref=ref,
+        )
+        .to_dataframe()
+        .reset_index()
+    )  # .to_csv('data_0.csv', index=False)
 
 
+def test_collection_properties(build_phases, test_table, obj):
+    ref = obj.ref
 
-def test_collection_properties(build_phases, test_table, ref):
     # for big builds, take advantage of progress bar, and parallel builds
     lnzs = np.linspace(-10, 3, 2000)
 
     # by default, progress bar hides itself after completion.  use context manager to keep it
     # note that for this example (where only have a single phase), doesn't really make a difference
-    with lnPi.set_options(tqdm_leave=True, joblib_use=False, tqdm_bar='text'):
-        o = lnPi.CollectionlnPi.from_builder(lnzs, build_phases)
+    with lnpy.set_options(tqdm_leave=True, joblib_use=False, tqdm_bar="text"):
+        o = lnpy.lnPiCollection.from_builder(lnzs, build_phases)
 
     other = get_test_table(o, ref)
 
@@ -132,49 +169,59 @@ def test_table_can():
 
 
 def get_test_table_can(ref):
-    return ref.xce.table(keys=['S','Z','betaE','betaF','betaOmega','betamu','dens','ntot']).to_dataframe().reset_index()
+    return (
+        ref.xce.table(
+            keys=["S", "Z", "betaE", "betaF", "betaOmega", "betamu", "dens", "ntot"]
+        )
+        .to_dataframe()
+        .reset_index()
+    )
 
-def test_canonical_properties(ref, test_table_can):
+
+def test_canonical_properties(obj, test_table_can):
+    ref = obj.ref
     other = get_test_table_can(ref)
     pd.testing.assert_frame_equal(other, test_table_can)
 
 
+def test_nice_grid(obj):
+    ref = obj.ref
+    build_phases = obj.build_phases
 
-def test_nice_grid(build_phases, ref):
-    import lnPi.collectionlnpiutils
+    import lnpy.lnpicollectionutils
 
-    with lnPi.set_options(joblib_use=True):
-        o_course, o = lnPi.collectionlnpiutils.limited_collection(build_phases, dlnz=0.01, offsets=[-10, +10],
-                                                                  even_grid=True, # but lnzs on same grid as dlnz
-                                                                  digits=2, #round lnzs to this number of digits
-                                                                  edge_distance_min=10, dens_min=0.001)
+    with lnpy.set_options(joblib_use=True):
+        o_course, o = lnpy.lnpicollectionutils.limited_collection(
+            build_phases,
+            dlnz=0.01,
+            offsets=[-10, +10],
+            even_grid=True,  # but lnzs on same grid as dlnz
+            digits=2,  # round lnzs to this number of digits
+            edge_distance_min=10,
+            dens_min=0.001,
+        )
 
     other_course = get_test_table(o_course, ref)
-    test_course = pd.read_csv(path_data / 'data_1_course.csv')
+    test_course = pd.read_csv(path_data / "data_1_course.csv")
     pd.testing.assert_frame_equal(other_course, test_course)
 
-
     other_fine = get_test_table(o, ref)
-    test = pd.read_csv(path_data / 'data_1_fine.csv')
+    test = pd.read_csv(path_data / "data_1_fine.csv")
     pd.testing.assert_frame_equal(other_fine, test)
-
 
     # spinodal/binodal
     o_course.spinodal(2, build_phases)
     o_course.binodal(2, build_phases)
 
-
     other = get_test_table(o_course.spinodal.access, ref)
-    test = pd.read_csv(path_data / 'data_1_spin.csv')
+    test = pd.read_csv(path_data / "data_1_spin.csv")
     pd.testing.assert_frame_equal(other, test)
-
 
     other = get_test_table(o_course.binodal.access, ref)
-    test = pd.read_csv(path_data / 'data_1_bino.csv')
+    test = pd.read_csv(path_data / "data_1_bino.csv")
     pd.testing.assert_frame_equal(other, test)
 
-
     # check dw
-    other = o_course.spinodal.access.wlnPi.dw.to_frame().reset_index()
-    test = pd.read_csv(path_data / 'data_1_spin_dw.csv')
+    other = o_course.spinodal.access.wfe.dw.to_frame().reset_index()
+    test = pd.read_csv(path_data / "data_1_spin_dw.csv")
     pd.testing.assert_frame_equal(other, test)
